# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OpenApi Documentation",
        "description": "Документация для CDP сервиса",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Основной API"
        }
    ],
    "paths": {
...
        "api/lk/bonuses/history": {
            "get": {
                "tags": [
                    "bonuses"
                ],
                "summary": "Получение бонусных транзакций пользователя",
                "description": "Защищено bearer-токеном",
                "operationId": "bonusesUserHistory",
                "parameters": [
                    {
                        "name": "bearer",
                        "in": "path",
                        "description": "bearer-токен пользователя",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "3|rePJ34DDFgds"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": "3"
                                            },
                                            "bonus_transaction": {
                                                "type": "string",
                                                "example": "+30"
                                            },
                                            "bonus_date": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "30.05.2001"
                                            },
                                            "bonus_id": {
                                                "type": "integer",
                                                "example": "3"
                                            },
                                            "bonus_type": {
                                                "type": "integer",
                                                "example": "3"
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": "3"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "api/lk/bonuses/balance": {
            "get": {
                "tags": [
                    "bonuses"
                ],
                "summary": "Получение баланса пользователя",
                "description": "Защищено bearer-токеном",
                "operationId": "bonusesUserBalance",
                "parameters": [
                    {
                        "name": "bearer",
                        "in": "path",
                        "description": "bearer-токен пользователя",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "3|rePJ34DDFgds"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": "3"
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": "30"
                                            },
                                            "balance": {
                                                "type": "integer",
                                                "example": "3000"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/lk/order/index": {
            "get": {
                "tags": [
                    "order"
                ],
                "summary": "Получение информации о купленных купонов пользователя",
                "description": "Защищено bearer-токеном",
                "operationId": "orderUserHistory",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "product_id": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    },
                                                    "product_count": {
                                                        "type": "integer",
                                                        "example": "100"
                                                    },
                                                    "product_duration": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "122"
                                                    },
                                                    "product_coupon": {
                                                        "type": "integer",
                                                        "example": "201"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/lk/order/create": {
            "post": {
                "tags": [
                    "order"
                ],
                "summary": "Создание заказа",
                "description": "Защищено bearer-токеном",
                "operationId": "createOrder",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "description": "id подарка",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "10"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "order_id": {
                                                        "type": "integer"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "scheme": "Bearer"
            }
        }
    }
}